refactor: tests and examples
