+++ media_types/extension_test.ts
@@ -1,0 +1,19 @@
+// Copyright 2018-2022 the Deno authors. All rights reserved. MIT license.
+
+import { assertEquals } from "../testing/asserts.ts";
+import { extension } from "./mod.ts";
+
+Deno.test({
+  name: "media_types - extension()",
+  fn() {
+    const fixtures: [string, string | undefined][] = [
+      ["image/gif", "gif"],
+      ["application/javascript", "js"],
+      ["text/html; charset=UTF-8", "html"],
+      ["application/foo", undefined],
+    ];
+    for (const [fixture, expected] of fixtures) {
+      assertEquals(extension(fixture), expected);
+    }
+  },
+});
