+++ media_types/get_charset_test.ts
@@ -1,0 +1,20 @@
+// Copyright 2018-2022 the Deno authors. All rights reserved. MIT license.
+
+import { assertEquals } from "../testing/asserts.ts";
+import { getCharset } from "./mod.ts";
+
+Deno.test({
+  name: "media-types - getCharset()",
+  fn() {
+    const fixtures = [
+      ["text/plain", "UTF-8"],
+      ["text/html", "UTF-8"],
+      ["application/foo", undefined],
+      ["application/news-checkgroups", "US-ASCII"],
+      ["application/news-checkgroups; charset=UTF-8", "UTF-8"],
+    ] as const;
+    for (const [fixture, expected] of fixtures) {
+      assertEquals(getCharset(fixture), expected);
+    }
+  },
+});
