+++ media_types/parse_media_type.ts
@@ -1,3 +1,4 @@
+// Copyright 2018-2022 the Deno authors. All rights reserved. MIT license.
 import { consumeMediaParam, decode2331Encoding } from "./_util.ts";
 
 /**
@@ -16,7 +17,7 @@ import { consumeMediaParam, decode2331Encoding } from "./_util.ts";
  *
  * @example
  * ```ts
- * import { parseMediaType } from "https://deno.land/std@$STD_VERSION/media_types/mod.ts";
+ * import { parseMediaType } from "https://deno.land/std@$STD_VERSION/media_types/parse_media_type.ts";
  * import { assertEquals } from "https://deno.land/std@$STD_VERSION/testing/asserts.ts";
  *
  * assertEquals(
