+++ media_types/type_by_extension.ts
@@ -1,3 +1,4 @@
+// Copyright 2018-2022 the Deno authors. All rights reserved. MIT license.
 import { types } from "./_db.ts";
 
 /**
@@ -8,7 +9,7 @@ import { types } from "./_db.ts";
  *
  * @example
  * ```ts
- * import { typeByExtension } from "https://deno.land/std@$STD_VERSION/media_types/mod.ts";
+ * import { typeByExtension } from "https://deno.land/std@$STD_VERSION/media_types/type_by_extension.ts";
  *
  * typeByExtension("js"); // `application/json`
  * typeByExtension(".HTML"); // `text/html`
