+++ media_types/type_by_extension_test.ts
@@ -1,0 +1,24 @@
+// Copyright 2018-2022 the Deno authors. All rights reserved. MIT license.
+
+import { assertEquals } from "../testing/asserts.ts";
+import { typeByExtension } from "./mod.ts";
+
+Deno.test({
+  name: "media_types - typeByExtension",
+  fn() {
+    const fixtures = [
+      ["js", "application/javascript"],
+      [".js", "application/javascript"],
+      ["Js", "application/javascript"],
+      ["html", "text/html"],
+      [".html", "text/html"],
+      [".HTML", "text/html"],
+      ["file.json", undefined],
+      ["foo", undefined],
+      [".foo", undefined],
+    ] as const;
+    for (const [fixture, expected] of fixtures) {
+      assertEquals(typeByExtension(fixture), expected);
+    }
+  },
+});
